Refactor method plotPoint.

+++ src/canvas.scala
@@ -444,11 +444,17 @@ class DrawCtx(comp: java.awt.Component, offs: Int = 0){
   def plotPoint(x: Double, y: Double) = (g: G2D) => {
     val radius = this.pointMarkSize
     val (xx, yy) = this.coordRangeToScreen((x, y))
-    g.drawString(s"(${x}, ${y})", xx, yy)
+    g.drawString(s"(${x}, ${y})", xx.toFloat, yy.toFloat)
 
     val col = g.getColor()
     g.setColor(this.pointMarkColor)
-    g.fillOval(xx - radius, yy - radius, 2 * radius, 2 * radius)
+    //g.fillOval(xx - radius, yy - radius, 2 * radius, 2 * radius)
+    g.fillOval(
+      (xx - radius).toInt,
+      (yy - radius).toInt,
+      (2 * radius).toInt,
+      (2 * radius).toInt
+    )
     g.setColor(col)
   }
